requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
openai==1.3.0
pypdf==3.17.0
python-docx==1.1.0
tiktoken>=0.8.0; python_version < '3.13'
pydantic-settings==2.1.0
pydantic==2.5.0

# Note: tiktoken is optional. If installation fails, the app will use character-based chunking as fallback.
# To install tiktoken manually: pip install tiktoken

